Add .gitignore & fix makefile & cast size_t to int in main.

+++ main.c
@@ -12,7 +12,7 @@ int main(int argc, const char * argv[]) {
     BIGNUM *bn = bn_new();
     bn_str2bn(bn, "012340567890");
 
-    printf("length of bn is %ld\n", bn_len(bn));
+    printf("length of bn is %d\n", (int)bn_len(bn));
     printf("content of bn is ");
     bn_print(stdout, bn);
